refactor: Makie general session and sessions pool
