+++ src/__tests__/scan-query.test.ts
@@ -7,10 +7,10 @@ import {
     initDriver,
     Row,
 } from '../test-utils';
-import {Session} from '../table';
+import {TableSession} from '../table/table-session';
 import {TypedData} from '../types';
 
-async function executeScanQuery(session: Session): Promise<TypedData[]> {
+async function executeScanQuery(session: TableSession): Promise<TypedData[]> {
     const query = `SELECT * FROM ${TABLE};`;
 
     const rows: TypedData[] = [];
