+++ src/table/session-event.ts
@@ -1,0 +1,4 @@
+export enum SessionEvent {
+    SESSION_RELEASE = 'SESSION_RELEASE',
+    SESSION_BROKEN = 'SESSION_BROKEN'
+}
